Combine various small common/ headers into common.h
There is no need to have separate files for just a define or two.
