+++ src/libhirte/bus/peer-bus.h
@@ -1,9 +1,8 @@
 #pragma once
 
 #include <netinet/in.h>
-#include <systemd/sd-bus.h>
 
-#include "libhirte/common/list.h"
+#include "libhirte/common/common.h"
 
 char *assemble_tcp_address(const struct sockaddr_in *addr);
