+++ src/libhirte/common/common.h
@@ -1,3 +1,51 @@
 #pragma once
 
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+
+#include <systemd/sd-bus.h>
+#include <systemd/sd-event.h>
+
+#include "list.h"
+
+#define streq(a, b) (strcmp((a), (b)) == 0)
+#define strneq(a, b, n) (strncmp((a), (b), (n)) == 0)
+
 #define UNUSED __attribute__((unused))
+
+static inline void *steal_pointer(void *ptr) {
+        void **ptrp = (void **) ptr;
+        void *p = *ptrp;
+        *ptrp = NULL;
+        return p;
+}
+
+static inline int steal_fd(int *fdp) {
+        int fd = *fdp;
+        *fdp = -1;
+        return fd;
+}
+
+static inline void closep(const int *fd) {
+        if (*fd >= 0) {
+                close(*fd);
+        }
+}
+
+static inline void freep(void *p) {
+        free(*(void **) p);
+}
+
+#define malloc0(n) (calloc(1, (n) ?: 1))
+
+#define _cleanup_(x) __attribute__((__cleanup__(x)))
+#define _cleanup_free_ _cleanup_(freep)
+#define _cleanup_fd_ _cleanup_(closep)
+
+#define _cleanup_sd_event_ _cleanup_(sd_event_unrefp)
+#define _cleanup_sd_event_source_ _cleanup_(sd_event_source_unrefp)
+#define _cleanup_sd_bus_ _cleanup_(sd_bus_unrefp)
+#define _cleanup_sd_bus_slot_ _cleanup_(sd_bus_slot_unrefp)
+#define _cleanup_sd_bus_message_ _cleanup_(sd_bus_message_unrefp)
+#define _cleanup_sd_bus_error_ _cleanup_(sd_bus_error_free)
