--- src/libhirte/common/strings.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <string.h>
-
-#define streq(a, b) (strcmp((a), (b)) == 0)
-#define strneq(a, b, n) (strncmp((a), (b), (n)) == 0)
