+++ src/libhirte/meson.build
@@ -6,13 +6,10 @@ libhirte_src = [
     'bus/user-bus.c',
     'bus/user-bus.h',
     'common/common.h',
-    'common/dbus.h',
     'common/list.h',
-    'common/memory.h',
     'common/opt.h',
     'common/parse-util.c',
     'common/parse-util.h',
-    'common/strings.h',
     'hashmap/hashmap.c',
     'hashmap/hashmap.h',
     'ini/config.c',
